+++ natcap/versioner/__init__.py
@@ -34,6 +34,31 @@ def get_version(package, root='.', ver_module=None):
 
     return vcs_version(root)
 
+def parse_version(root='.'):
+    """
+    Determine the correct source from which to parse the version.
+
+    If PKG-INFO exists, then we're in a source or binary distribution so prefer
+    to extract this metadata first.  Otherwise, If we're in an hg or git repo,
+    get the version from SCM.
+
+    Parameters:
+        root='.' (string): The root directory to search for vcs information.
+            This should be the path to the repository root.
+
+    Returns:
+        A versioning string.
+    """
+
+    pkginfo_filepath = os.path.join(root, 'PKG-INFO')
+    if os.path.exists(pkginfo_filepath):
+        with open(pkginfo_filepath) as pkginfo_file:
+            for line in pkginfo_file:
+                if line.startswith('Version'):
+                    return line.split(': ')[1].rstrip()
+
+    return vcs_version(root)
+
 def vcs_version(root='.'):
     """
     Get the version string from your VCS.